+++ src/Vending/IVendingSession.cs
@@ -3,6 +3,6 @@
     public interface IVendingSession
     {
         bool TryAcceptToken(Token token);
-        double GetCurrentCoinValue();
+        decimal GetCurrentCoinValue();
     }
 }
